Rework benchmark
The old benchmark resulted in a 1ns iteration time which basically means
that it wasn't run or optimized away. I tried to wrap it in a
`test::black_box` and to force the return value assignment but it does
not change anything so rewrote the benchmark to check how long it takes
to read and write 1024 32-bit floats in a buffer.